refactor(oop_in_action): change the scene name
- Separate the shared components
