+++ animations/oop_in_action/src/scenes/class_object.meta
