+++ animations/oop_in_action/src/shared/styles.ts
