+++ shared/animations/styles.ts
